Initializing Pokemon Smash 1.0
This is the first stable version of Pokemon Smash. This contains the framework within the file custom_library.py, which details the Pokemon and Move classes. This enables the game to be largely scalable. Once I built the foundation, it was just a matter of creating the gameplay logic and arithmetic, as well as adding in the featured pokemon and their respective moves.